Support proposal deposits in single choice voting module.
This adds a `deposit_info` field to the voting module. In that field
one may specify information about the address of the token that
deposits should be taken from, the amount of tokens to deposit, and if
failed proposals should be refunded.

Optionally, instead of specifying the specific address of a token a
module may specify that it would like to use whatever token that
voting module is using. To support this the voting module must have a
`TokenContract {}` query message which may be derived with a new
`token_info` procedural macro.

Specifying that the token address should be the voting module's token
address will cause instantiation to fail in the event that that voting
module does not implement the `TokenContract {}` query.

Deposits are refunded and deposited the same way as the existing
cw3-dao handles things.