Update contract naming scheme, and cleanup docs.
This updates the naming scheme for DAO DAO contracts. All contracts
are given the prefix `cw-dao`. The core module is renamed to be
`cw-dao-core`. Proposal modules are named in the format
`cw-dao-proposal-<proposal_type>` and voting modules are named in the
format `cw-dao-voting-<voting_type>`.

This also improves the README situation for many contracts.

Possibly controversially, this removes the `cw-named-groups` and
`cw-names-registry` contracts. The thinking being: our design
direction has changed since those were written. For the names
registry, I'm currently imagining that we will build out a much more
sophisticated system here when we do vote delegation that will replace
this contract. For named groups, we now support strings in
`cw-dao-core` items. This probably replaces the need for this contract
as frontends will likely just want to store this information in that
string.