More information pointing to a dead link
It looks like https://github.com/DA0-DA0/dao-contracts/wiki/DAO-DAO-v1-Contracts-Design is no longer a valid link. 

Shall we delete the following? 'For more information about how these modules fit together see
[this](https://github.com/DA0-DA0/dao-contracts/wiki/DAO-DAO-v1-Contracts-Design)
wiki page.'

Or is there information in the wiki we can update to

Or do we need to create information in the wiki about the cw-core contract.