Add `NextProposalId` query to proposal module interface.
Doing so pretty significantly simplifies pre-propose modules, and
makes less compex to write as we can remove the proposal created hook
entirely as part of this change.

This also walks back the addition of pre-propose and hooks to the
proposal module interface. I think there is beauty in smallness there,
and if someone else implements a proposal module they may not want to
use our hook and proposal-gating types.