More informative error to address BlockBeforeSend hook executing on minting or burning
This is intended functionality. If a token is frozen, a DAO needs to
both grant a minter allowance as well as adding the minter to the
allowlist to allow for token transfers when the token is frozen.

The new error message should make next steps clear if this edge case is
encountered.